pyproject.toml
[project]
name = "task-manager-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "authlib>=1.3.2",
    "fastapi>=0.115.5",
    "mypy>=1.13.0",
    "psycopg[binary]>=3.2.3",
    "pydantic>=2.9.2",
    "pyjwt>=2.10.0",
    "python-gnupg>=0.5.3",
    "ruff>=0.8.0",
    "uvicorn[standard]>=0.32.0",
]
